requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos[dbt-bigquery]==1.3.1
protobuf==3.20.0
pandas